Add files via upload
Try to wrap the api service into a container , and connect with the other container which has a PostgreSQL DB in it.
1. Success to build a docker image of our api service, however, the container it produced has difficulty connected to the other container with PostgreSQL DB.
(I think that the issues here are that we are not able to get a right connection between two containers, e.g. IP and its PORT  are different )
Though two containers can be linked by command "--link", while running the image as a container.